% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{choose_chunk_size}
\alias{choose_chunk_size}
\title{Select a reasonable chunk size for an HDF5 dataset object}
\usage{
choose_chunk_size(x)
}
\arguments{
\item{x}{A vector to store in an HDF5 file}
}
\value{
a numeric value containing a suggested chunk size
}
\description{
Select a reasonable chunk size for an HDF5 dataset object
}
