% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{merge_h5_seurat}
\alias{merge_h5_seurat}
\title{Reads a list of h5 files, merges them together and returns a single Seurat Object}
\usage{
merge_h5_seurat(h5_path_list)
}
\arguments{
\item{h5_path_list}{list of filepaths to h5 files.}
}
\value{
a Seurat Class object
}
\description{
By default, matrix data will be stored in the "RNA" assay.
}
