% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_aligned_barplot}
\alias{qc_aligned_barplot}
\title{Generate a baseline-aligned barplot for two categorical metrics}
\usage{
qc_aligned_barplot(
  meta,
  category_x = "batch_id",
  name_x = "Batch ID",
  category_y = "well_id",
  category_name = "Well ID",
  colorset_y = "varibow",
  name_y = "N Cells",
  padding = 0.2
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{category_x}{A character object specifying the metadata to use for grouping on the x-axis}

\item{name_x}{A character object specifying a name to display on the x-axis}

\item{category_y}{A character object specifying the metadata to use for splitting in the y-direction}

\item{category_name}{A character object specifying a name to display for the colors}

\item{colorset_y}{A colorset to use as fills for category_y. Currently supported: "rainbow" or "varibow". Default is "varibow"}

\item{name_y}{A character object specifying a name for the y-axis.}

\item{padding}{A numeric object specifying the fraction of the total vertical space to use for separating category_y groups. Default is 0.2.}
}
\value{
a ggplot2 plot object
}
\description{
The metric used for category_x will generate bars as columns on the x-axis.
The bar will be split vertically based on category_y. Each group in category_y will be aligned to make them easier to compare.
}
