% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{read_h5_seurat}
\alias{read_h5_seurat}
\title{Read .h5 directly into a Seurat Object}
\usage{
read_h5_seurat(h5_file, target = "matrix", feature_names = "name", ...)
}
\arguments{
\item{h5_file}{the path to an .h5 file in 10x Genomics format}

\item{target}{A matrix object in the .h5 file with a /features/ sub-group. Default is "matrix".}

\item{feature_names}{a character object specifying whether to use "id" or "name" for row.names. Default is "name".}

\item{...}{Additional parameters passed to \code{\link[SeuratObject]{createSeuratObject}}}
}
\value{
a Seurat Class object
}
\description{
By default, matrix data will be stored in the "RNA" assay.
}
\details{
CITE-seq data will be automatically detected and stored in the "ADT" assay.
}
\examples{
h5_fpath <- system.file('testdata/cite_sample1.h5', package = "H5weaver")
so <- read_h5_seurat(h5_fpath)
# so
}
