% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_h5.R
\name{read_h5_vector}
\alias{read_h5_vector}
\title{Read 1d arrays from h5 files as vectors}
\usage{
read_h5_vector(...)
}
\arguments{
\item{...}{Passes arguments to `rhdf5::h5read()`}
}
\value{
a vector object
}
\description{
This function wraps `h5read()` with `as.vector()` so that 1d arrays are 
converted to vectors so they behave as expected in R.
}
