% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5ad.R
\name{read_h5ad_feature_meta}
\alias{read_h5ad_feature_meta}
\title{Read .h5ad Feature Metadata}
\usage{
read_h5ad_feature_meta(h5ad_file)
}
\arguments{
\item{h5ad_file}{the path to an .h5ad file}
}
\value{
a data.frame containing all feature metadata found in /var
}
\description{
Read .h5ad Feature Metadata
}
