% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{split_h5_by_hash}
\alias{split_h5_by_hash}
\title{Split a 10x HDF5 file based on HTOparser results}
\usage{
split_h5_by_hash(h5_file, hash_category_table, hash_count_matrix = NULL,
  add_uuid = TRUE, add_name = TRUE)
}
\arguments{
\item{h5_file}{a character object specifying the location of a 10x Genomics .h5 file.}

\item{hash_category_table}{a data.frame or data.table with results from HTOparser::categorize_binary_hash_matrix() consisting of 3 columns: cell_barcode, hto_category, and hto_barcode.}

\item{hash_count_matrix}{(optional) a matrix or dgCMatrix with results from CITE-seq-Count, HTOparser::add_missing_hto_rows(), or HTOparser::barcode_table_to_matrix().}
}
\value{
A list object with contents of the input h5_file separated by each hto_barcode. Also includes multiplets.
}
\description{
Split a 10x HDF5 file based on HTOparser results
}
