% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rna_h5.R
\name{subset_h5_list_by_barcodes}
\alias{subset_h5_list_by_barcodes}
\title{Subset a h5_list object based on a set of barcodes.}
\usage{
subset_h5_list_by_barcodes(h5_list, barcodes)
}
\arguments{
\item{h5_list}{a list object generated by running rhdf5::h5dump() on a 10x HDF5 file, with a matrix converted by h5_list_convert_to_dgCMatrix().}

\item{barcodes}{a set of barcodes to select for filtering}
}
\value{
a list object
}
\description{
Subset a h5_list object based on a set of barcodes.
}
