% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{write_10x_h5_container}
\alias{write_10x_h5_container}
\title{Generate an empty HDF5 container to store RNA-seq count data in 10x format}
\usage{
write_10x_h5_container(
  h5_file,
  cell_barcode_length = 15,
  data_type = "integer",
  data_bits = 16,
  indices_bits = 16,
  indptr_bits = 32
)
}
\arguments{
\item{h5_file}{A character object specifying the location to write the HDF5 file. Required.}

\item{cell_barcode_length}{Integer specifying the cell barcode length. Default is 15.}

\item{data_type}{String specifying the type of the data to be written to the matrix. Default is "integer".}

\item{data_bits}{Integer specifying the bits to use to store matrix data. Default is 16.}

\item{indices_bits}{Integer specifying the bits to use to store indices. Default is 16.}

\item{indptr_bits}{Integer specifying the bits to use to store index pointers. Default is 32.}
}
\description{
Generate an empty HDF5 container to store RNA-seq count data in 10x format
}
