% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{write_ext_h5_character}
\alias{write_ext_h5_character}
\title{Write values to an extensible charachter HDF5 dataset}
\usage{
write_ext_h5_character(object, h5_handle, name)
}
\arguments{
\item{name}{}
}
\description{
If the dataset
}
