% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rna_h5.R
\name{write_seurat_h5}
\alias{write_seurat_h5}
\title{Write a Seurat Object to an .h5 file}
\usage{
write_seurat_h5(so, h5_file, overwrite = FALSE)
}
\arguments{
\item{so}{a Seurat Object}

\item{h5_file}{a character object specifying the location of a .h5 file to write to.}

\item{overwrite}{a logical value specifying whether or not to overwrite an existing .h5 file. Default is FALSE.S}
}
\value{
Writes a file; no return to R.
}
\description{
Note: Only retains count data, not normalized values
}
